==> 01_Data_prep_v2_PN_1_Dec_2021.ipynb <==
# import libraries
import pandas as pd
import glob
import os
import numpy as np

# 1. Read and combine all csv files

os.chdir('./data/csv')
# Read all csv filenames and combine a list of them
filenames =  [i for i in glob.glob('*.{}'.format('csv'))]
# Sort filenames in list to ensure dates are in order
filenames = sorted(filenames)

# Read first file as a reference table and format features
df = pd.read_csv(filenames[0])
year = str(filenames[0]).split("_")[2]
quarter = str(filenames[0]).split("_")[3].replace('iv', '4').replace('iii', '3').replace('ii', '2').replace('i','1')
filenames.pop(0)
df['year_quarter'] = str(year)+"_"+str(quarter)
df['Registration_number'] = df['Registration_number'].astype(str)

df.info()
df.head()

# Read all other csv files and add relevant data to the initial dataframe
#for i in range (1, len(filenames)):
for f in filenames:
    print(f)
    # Extract period information
    year = str(f).split("_")[2]
    quarter = str(f).split("_")[3].replace('iv', '4').replace('iii', '3').replace('ii', '2').replace('i','1')
    df_new = pd.read_csv(f)
    df_new['year_quarter'] = str(year)+"_"+str(quarter)
    df_new['Registration_number'] = df_new['Registration_number'].astype(str)
    df.reset_index(drop=True)
    df = pd.concat([df, df_new])

df.info()
df.head()


2. Read data from Statistics Estonia provided Excel peradsheets and combine with our database

os.chdir('..')
os.chdir('..')
os.chdir('./data/stat')

# Read all csv filenames and combine a list of them
filenames =  [i for i in glob.glob('*.{}'.format('xlsx'))]
# Sort filenames in list in reverse order to ensure dates are in order and newer data gets used first
filenames.sort(reverse=True)
filenames

# Read first file as a reference table and format features
df_stat = pd.read_excel(filenames[0])
df_stat = df_stat.rename(columns={  'Registrikood': 'Registration_number',\
                                    'Nimi': 'Name', \
                                    'EMTAK2008 Tähtkoodiga valdkond': 'EMTAK_by_letter',\
                                    'EMTAK2008 4-kohaline kood': 'EMTAK_by_number',\
                                    'Institutsionaalsete sektorite Eesti jaoks kohandatud klassifikaator 2010': 'Institutional_classification' ,\
                                    'Kliendi/emitendi identifikaator (vastavalt Eesti Panga presidendi 29. mai määrusele nr 7)': 'Client_ID_by_Eesti_Pank',\
                                    'Üksuse registreerimise kuupäev': 'Registration_date',\
                                    'Üksuse registrist kustutamise kuupäev' : 'Deletion_Date'})

# Remove the loaded filename from files list
filenames.pop(0)
df_stat = df_stat.iloc[: , :-1]
df_stat['Registration_number'] = df_stat['Registration_number'].astype(str)

df_stat.info()
df_stat.head()

# Add other statistical data, only that is missing from previous
for f in filenames:
    print(f)
    df_stat_new = pd.read_excel(f)
    df_stat_new = df_stat_new.rename(columns={  'Registrikood': 'Registration_number',\
                                                'Nimi': 'Name', \
                                                'EMTAK2008 Tähtkoodiga valdkond': 'EMTAK_by_letter',\
                                                'EMTAK2008 4-kohaline kood': 'EMTAK_by_number',\
                                                'Institutsionaalsete sektorite Eesti jaoks kohandatud klassifikaator 2010': 'Institutional_classification' ,\
                                                'Kliendi/emitendi identifikaator (vastavalt Eesti Panga presidendi 29. mai määrusele nr 7)': 'Client_ID_by_Eesti_Pank',\
                                                'Üksuse registreerimise kuupäev': 'Registration_date',\
                                                'Üksuse registrist kustutamise kuupäev' : 'Deletion_Date'})
    df_stat_new['Registration_number'] = df_stat_new['Registration_number'].astype(str)
    df_stat = pd.concat([df_stat, df_stat_new], axis=0)

# Dropping duplicate entries
df_stat = df_stat.drop_duplicates(subset=['Registration_number'])
df_stat.reset_index(drop=True)
df_stat.head()

#df.reset_index(drop=True)
#df_stat.reset_index(drop=True)

# Combining 2 datasets into one large
# Loose name column
df_stat.drop(df_stat.columns[[1]], axis = 1, inplace = True)

df_merged = pd.merge(left= df, right=df_stat, how='outer', left_on= 'Registration_number', right_on='Registration_number')
df_merged.head()

df_merged.info()
df_merged.head()

df_merged.shape
#df_merged.tail()


# Period-wise
We have considered 3 different time-frames for testing our model:
- before COVID-19 (2017 ... 1st quarter of 2020)
- during COVID-19 (2nd quarter of 2020 ... 3rd quarter 2021)
- the whole period

The idea is to see weather and how COVID-19 pandemic has affected different companies. Therefore we create 2 additional datasets

df_merged.nunique()

df_merged['year_quarter'].unique().tolist()

# Preliminary filtering - all companies that did not have data from taxes datasets will be discarded

df_merged = df_merged[df_merged['year_quarter'].notna()]

print(df_merged.shape)
df_merged['year_quarter'].unique().tolist()

# Save acquired data into a csv

os.chdir('..')
os.chdir('..')
os.chdir('./data/dataframes/')
df_merged.to_csv( "full_period.csv", index=False)

# Filter out data that is after 1st quarter of 2020

df_before = df_merged.copy(deep=True)

df_before = df_before.loc[df_before.apply(lambda x: x.year_quarter in ['2017_1', '2017_2', '2017_3', '2017_4', '2018_1', '2018_2', '2018_3', '2018_4', '2019_1', '2019_2', '2019_3', '2019_4', '2020_1'], axis=1)]

df_before.head()

df_before.shape

# Save acquired data into a csv

df_before.to_csv( "before_covid.csv", index=False)

# Filter out data that is since 2nd quarter of 2020

df_during = df_merged.copy(deep=True)

df_during = df_during.loc[df_during.apply(lambda x: x.year_quarter in ['2020_2', '2020_3', '2020_4', '2021_1', '2021_2', '2021_3'], axis=1)]


df_during.head()

df_during.shape

# Save acquired data into a csv

df_during.to_csv( "during_covid.csv", index=False)
