==> 02_Data_cleaning_V1_PN_1_Dec_2021.ipynb <==
# import necessary libraries
import pandas as pd
import numpy as np
import math
import os

# Data Cleaning

# Loading all 3 databases

all_df = pd.read_csv('./data/dataframes/full_period.csv')
all_df = all_df.sort_values('year_quarter').reset_index(drop=True)

before_df = pd.read_csv('./data/dataframes/before_covid.csv')
before_df = before_df.sort_values('year_quarter').reset_index(drop=True)

during_df = pd.read_csv('./data/dataframes/during_covid.csv')
during_df = during_df.sort_values('year_quarter').reset_index(drop=True)

all_df.shape

before_df.shape

during_df.shape

# Define data cleaning function

def clean_data(df):
    all_df = df

    # drop turnover missing values
    all_df = all_df[all_df['Turnover'].notna()]

    # Remove all business types except äriühing
    all_df = all_df[all_df['Type']=="Äriühing"]

    # Drop duplicates
    all_df = all_df.drop_duplicates()

    # impute missing values
    all_df['County'] = all_df['County'].replace(np.nan, "other")
    all_df['Business_area_by_EMTAK_system'] = all_df['Business_area_by_EMTAK_system'].fillna("other")

    # same as number of employee, will do it by grouping during feat eng
    #all_df['All_taxes'] = all_df['All_taxes'].fillna(all_df['All_taxes'].mean())
    #all_df['Labor_taxes'] = all_df['Labor_taxes'].fillna(all_df['Labor_taxes'].mean())

    # will add this to the feat engineering
    # to do it by grouping
    #all_df['number_of_employees'] = all_df['number_of_employees'].fillna(int(all_df['number_of_employees'].mean()))

    # nan values mean companu still exist
    all_df['Deletion_Date']=all_df['Deletion_Date'].fillna(int(1))
    # Other values mean company died
    all_df.loc[(all_df['Deletion_Date'] !=1), 'Deletion_Date'] = 0

    #replacing negative turover with small nr
    all_df['Turnover'] = all_df['Turnover'].apply(lambda x : x if x > 0 else 0.000001)

    #dropping name_y
    #all_df.drop(columns=['Name_y'], inplace=True)

    return all_df

# Cleaning datasets

all_cl = clean_data(all_df).reset_index(drop=True)
before_cl = clean_data(before_df).reset_index(drop=True)
during_cl = clean_data(during_df).reset_index(drop=True)
#all_cl

all_cl.shape

before_cl.shape

during_cl.shape

# Testing on company Nortal AS
print('Whole period')
print(all_cl.query('Registration_number == 10391131').shape)
print(all_df.query('Registration_number == 10391131').shape)

print('Before covid')
print(before_cl.query('Registration_number == 10391131').shape)
print(before_df.query('Registration_number == 10391131').shape)

print('During covid')
print(during_cl.query('Registration_number == 10391131').shape)
print(during_df.query('Registration_number == 10391131').shape)

# Feature Engineering

# Defining engineering engine

def feat_eng(g):
    ex_df=pd.DataFrame(g)

    if ex_df['number_of_employees'].isnull().sum()>0:

        try:
            ex_df['number_of_employees'] = ex_df['number_of_employees'].fillna(int(ex_df['number_of_employees'].mean()))
        except:
            ex_df['number_of_employees'] = ex_df['number_of_employees'].fillna(int(1))



    if ex_df['Labor_taxes'].isnull().sum():
        #print("Labor_taxes")
        if math.isnan(ex_df['Labor_taxes'].mean()):
            ex_df['Labor_taxes'] = 0000.1
        else:
            ex_df['Labor_taxes'] = ex_df['Labor_taxes'].fillna(ex_df['Labor_taxes'].mean())


    if ex_df['All_taxes'].isnull().sum():
        if math.isnan(ex_df['All_taxes'].mean()):
            ex_df['All_taxes'] = 0000.1
        else:
            ex_df['All_taxes'] = ex_df['All_taxes'].fillna(ex_df['All_taxes'].mean())



    # Cumulative Turn over
    ex_df['cum_sum_turnover'] = np.round(ex_df['Turnover'].cumsum(), 2)
    ex_df['cum_perc_turnover'] = np.round(100*ex_df['cum_sum_turnover']/ex_df['Turnover'].sum(), 2)

    # combined taxes
    ex_df['combined_taxes'] = np.round(ex_df['Labor_taxes'] + ex_df['All_taxes'], 2)

    # Cumulative Taxes
    ex_df['cum_sum_combined_taxes'] = np.round(ex_df['combined_taxes'].cumsum(), 2)
    #ex_df['cum_perc_combined_taxes'] = np.round(100*ex_df['cum_sum_combined_taxes']/ex_df['combined_taxes'].sum(), 2)

    # Revenue Growth
    ex_df['revenue_growth'] = np.round(ex_df.Turnover.pct_change(fill_method="ffill"),2)
    ex_df['revenue_growth'] = ex_df['revenue_growth'].fillna(0)


    # Profit per Employee
    ex_df["profit_per_employee"] = np.round(ex_df['Turnover']/ex_df['number_of_employees'], 2)


    return ex_df

# Doing magic on whole period data

all_cl = all_cl.groupby("Registration_number").apply(feat_eng)
all_cl

# Saving cleared and engineered data

all_cl.to_csv('./data/dataframes/full_period_enginered.csv', index=False)
# Just to add something

all_cl.query('Registration_number == 10391131')

# Not necessary as will be divided into parts after cleaning/engineering/labelling is complete
# ... with pre-covid period of data
'''
before_cl = before_cl.groupby("Registration_number").apply(feat_eng)
before_cl

# ... and with the period of pandemic

during_cl = during_cl.groupby("Registration_number").apply(feat_eng)
during_cl


before_cl.to_csv('./data/dataframes/before_covid_enginered.csv', index=False)

during_cl.to_csv('./data/dataframes/during_covid_enginered.csv', index=False)
'''
