==> 03_Data_engineering_P2_V1_PN_1_Dec_2021.ipynb <==
# Feature Engineering continuing

# Import libraries
import pandas as pd
from numpy import log as ln
import matplotlib.pyplot as plt
from numpy import inf

# Load data

df_cleaned = pd.read_csv('./data/dataframes/full_period_enginered.csv')
df_cleaned.head()

# Define function

def add_label_feat(g):
    ex_df = pd.DataFrame(g).reset_index()
    #print(ex_df)
    E_beg = list(ex_df['number_of_employees'])[0]
    E_end = list(ex_df['number_of_employees'])[-1]
    try:
        Birch_index = (E_end-E_end)*(E_end/E_beg)
    except:
        Birch_index = 0
    ex_df['Birch_index'] = Birch_index

    #print(ex_df['year_quarter'])

    st_year = int(list(ex_df['year_quarter'])[0].split("_")[0])
    st_year

    #print(list(ex_df['year_quarter']))
    st_quarter = int(list(ex_df['year_quarter'])[0].split("_")[1])
    #st_quarter = get_int_quarter(st_quarter)
    st_quarter

    en_year = int(list(ex_df['year_quarter'])[-1].split("_")[0])
    en_year


    en_quarter = int(list(ex_df['year_quarter'])[-1].split("_")[1])
    #en_quarter = get_int_quarter(en_quarter)
    en_quarter

    dur_q =(4 - (st_quarter - 1)) + (((en_year-st_year) - 1) *4) + en_quarter


    # employment_growth_rate
    emp_growth_rate_CGR = (ln(E_end) - ln(E_beg))/dur_q
    ex_df['emp_growth_rate_CGR'] = emp_growth_rate_CGR

    # absolute change in the number of employees
    ACE = abs(E_end - E_beg)
    ex_df['ACE'] = ACE


    return ex_df

df_cleaned = df_cleaned.groupby("Registration_number").apply(add_label_feat).reset_index(drop=True)
df_cleaned

df_cleaned.nunique()

# Correcting indexes

df_cleaned = df_cleaned.drop(["index"], 1).reset_index(drop=True)
df_cleaned

df_cleaned.query('Registration_number == 10391131')

# Savin data to file

df_cleaned.to_csv('./data/dataframes/full_period_4_labelling.csv')

# Visualising amount of data we have on companies (max 17 quarter data)

df_companies = df_cleaned.groupby('Registration_number').count()
df_companies = df_companies.iloc[:,0:1]
df_companies['Reg_no'] = range(1, len(df_companies) + 1)
df_companies['Reg_no'] = df_companies.index
df_companies.reset_index(drop=True, inplace=True)
df_companies.rename(columns={'Name':'Occurrence'}, inplace=True)
df_companies.head()

df_quarters = df_companies.groupby('Occurrence').count()
df_quarters['Quarters'] = range(1, len(df_quarters) + 1)
df_quarters['Quarters'] = df_quarters.index
df_quarters.reset_index(drop=True, inplace=True)
df_quarters.rename(columns={'Reg_no':'Times'}, inplace=True)
df_quarters

As we have vrey little refference point with the companies what have only 1 or few qarters of quality data, it would be wise to drop them
